Implement database connection and model management for Git analysis tool
- Added connection management for SQLAlchemy in `connection.py`, including initialization, session handling, and cleanup.
- Created database models in `model.py` for various entities such as commits, files, and analysis results.
- Implemented database operations in `operation.py` for inserting commits, diff results, commit files, analysis results, and file snapshots.
- Developed utility functions in `utils.py` for handling file metrics, tech stack identification, and CSV writing.
- Introduced wrappers in `wrapper.py` for analyzing snapshots and diffs, encapsulating logic for file processing and database interaction.
- Refactored main execution logic in `main.py` to utilize the new database and analysis structure, including argument parsing and execution flow.
