+++ git2base/analyzers/FileCharCountAnalyzer.py
