+++ git2base/analyzers/RegexMatchCountAnalyzer.py
