+++ git2base/analyzers/XMLElementCountAnalyzer.py
