+++ git2base/analyzers/base_analyzer.py
