+++ git2base/database/operation.py
@@ -1,5 +1,5 @@
-from database.connection import session_scope
-from database.model import AnalysisResult, Commit, CommitFile, DiffResult, FileSnapshot
+from git2base.database.connection import session_scope
+from git2base.database.model import AnalysisResult, Commit, CommitFile, DiffResult, FileSnapshot
 
 
 def insert_commits(commits_data: list[Commit]) -> list[Commit]:
